refactor: rename `packages` to `bindings`
