+++ bindings/go/constants/constants.go
