+++ bindings/go/go.mod
@@ -1,0 +1,3 @@
+module github.com/ossf/osv-schema/bindings/go
+
+go 1.22.7
