--- packages/go/go.mod
@@ -1,3 +1,0 @@
-module github.com/ossf/osv-schema/packages/go
-
-go 1.22.7
